Use TSD to improve logging.
* src/payprocd.c (my_tsd_key): New.
(pid_suffix_callback): New.
(main): Setup TSD.
(connection_thread): Store connection ID in the TSD.
* src/paypal-ipn.c (paypal_proc_ipn): Remove arg IDNO.
--

This allows us to print an identifier for the thread with the log
messages.  To further improve on that we may want to move that all to
logging and keep some logging specific information in the TSD.  For
example this could help to decide whether a LF needs to be printed if
a log message from a second thread is to be printed after an
incomplete log message from the first thread.