Implement an explicit test mode.
* configure.ac (PAYPROCD_TEST_SOCKET_NAME): New ac_define.
* src/payproc-post.c (opts): New options --live and --test.
(send_request): Use the test socket in test mode.  Print a warning if
neither --live nor --test are given.
* src/payprocd.c (opts): New option --test.
(main): Print a warning if neither --live nor --test are given.
(server_socket_name): New.
(launch_server, cleanup): Use new function.
--

Note that ppipnhd does not yet allow for this.  We need to look at the
received data to decide whether this was test or live data.

Also payproc-post now requires the use of --live because it makes more
sense to default to --test.

Signed-off-by: Werner Koch <wk@gnupg.org>