Do not fail a call to paypal or stripe with no return data.
* src/paypal.c (call_paypal): Replace an empty string by a JSON
object.
* src/stripe.c (call_stripe): Ditto.
--

Some Paypal call s do not return data.  This results in an empty
string in our functions which is a non-valid JSON object.  The effect
is that an error is returned and not an "empty" JSON object.  To
mitigate this we return the "null" JSON object in such a case.

Signed-off-by: Werner Koch <wk@gnupg.org>